Remove unneeded reductive rowspec databag generator
